drop Recommends: gimp-jxl-plugin, avoid pulling GIMP into WS
It's not a good idea for a library package to Recommend something
as app-specific as a GIMP plugin. Anything that uses libjxl will
pull in the whole of GIMP unless soft dependencies are disabled,
which pulls in Python 2.7 and GTK+ 2. Since aom added a dependency
on libjxl, this was actually affecting our Workstation image
composes.